Added redirection to the login and reg pages
This will allow users to click the links to redirect them to the
respective pages if they are yet to make an account, or have already
made one and need to login. Successfully logging in now directs you to
a homepage that can be edited to any other template or directory.