url, view, html changes
added logout function so that the user logged in is actually removed once you press the button,

added a decorative line to the dashboard that checks whether a user is logged in and redirects them to an 'access denied' page if not logged in.

added a couple more html pages to support the new view(s) and urls.

made some changes to settings .py to allow for boostrap4 usage.